optimize: continue optimize benchmark_5000_basic_boxes->create and dispose gameObjects->memory: now not increate memory: reason GameObjectSystem,ThreeDTransformSystem,MaterialSystem->xxxMap will not release memory when delete by set map[key] = void 0!
solution
use es6->Map instead of object
advantage:
the spend time of Map->delete is equal to Map.set

disadvantage:
Map.set,get is slower than object->set,get!

+++ src/component/transform/operateDataSystem.ts
@@ -97,7 +97,7 @@ var _changeMapData = (sourceIndex: number, targetIndex: number, changeFunc:(tran
 //     dataArr[index] = void 0;
 // }
 
-var _moveToTypeArr = (dataArr: Uint16Array, sourceIndex: number, targetIndex: number, length: number
+var _moveToTypeArr = (dataArr: Float32Array, sourceIndex: number, targetIndex: number, length: number
 ) => {
     for (let i = 0; i < length; i++) {
         let newIndex1 = sourceIndex + i,
